Rename to faim-ipa.
