+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2023, Friedrich Miescher Institute for Biomedical Research
+Copyright (c) 2024, Friedrich Miescher Institute for Biomedical Research
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
