+++ pyproject.toml
@@ -7,4 +7,4 @@ requires = [
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools_scm]
-write_to = "src/faim_hcs/_version.py"
+write_to = "src/faim_ipa/_version.py"
