+++ setup.cfg
@@ -1,9 +1,9 @@
 [metadata]
-name = faim_hcs
-description = Tools used at FMI-FAIM to handle HCS data.
+name = faim_ipa
+description = Tools used at FMI-FAIM for Image Processing and Analysis.
 long_description = file: README.md
 long_description_content_type = text/markdown
-url = https://github.com/fmi-faim/faim-hcs.git
+url = https://github.com/fmi-faim/faim-ipa.git
 author = Tim-Oliver Buchholz
 author_email = tim-oliver.buchholz@fmi.ch
 license = BSD-3-Clause
@@ -18,10 +18,10 @@ classifiers =
     Programming Language :: Python :: 3 :: Only
     Topic :: Software Development :: Testing
 project_urls =
-    Bug Tracker = https://github.com/fmi-faim/faim-hcs/issues
-    Documentation = https://github.com/fmi-faim/faim-hcs#README.md
-    Source Code = https://github.com/fmi-faim/faim-hcs
-    User Support = https://github.com/fmi-faim/faim-hcs/issues
+    Bug Tracker = https://github.com/fmi-faim/faim-ipa/issues
+    Documentation = https://github.com/fmi-faim/faim-ipa#README.md
+    Source Code = https://github.com/fmi-faim/faim-ipa
+    User Support = https://github.com/fmi-faim/faim-ipa/issues
 
 [options]
 packages = find:
