+++ src/faim_ipa/hcs/imagexpress/ImageXpressWellAcquisition.py
@@ -3,10 +3,10 @@ from typing import Optional, Union
 
 import pandas as pd
 
-from faim_hcs.hcs.acquisition import TileAlignmentOptions, WellAcquisition
-from faim_hcs.io.MetaSeriesTiff import load_metaseries_tiff_metadata
-from faim_hcs.stitching import Tile
-from faim_hcs.stitching.Tile import TilePosition
+from faim_ipa.hcs.acquisition import TileAlignmentOptions, WellAcquisition
+from faim_ipa.io.MetaSeriesTiff import load_metaseries_tiff_metadata
+from faim_ipa.stitching import Tile
+from faim_ipa.stitching.Tile import TilePosition
 
 
 class ImageXpressWellAcquisition(WellAcquisition):
