+++ src/faim_ipa/hcs/imagexpress/MixedAcquisition.py
@@ -1,23 +1,26 @@
 import re
-from decimal import Decimal
 from pathlib import Path
 from typing import Optional, Union
 
-import numpy as np
 import pandas as pd
+from numpy._typing import NDArray
 
-from faim_hcs.hcs.acquisition import TileAlignmentOptions
-from faim_hcs.hcs.imagexpress import ImageXpressPlateAcquisition
-from faim_hcs.io.MetaSeriesTiff import load_metaseries_tiff_metadata
+from faim_ipa.hcs.acquisition import TileAlignmentOptions
+from faim_ipa.hcs.imagexpress import StackAcquisition
 
 
-class StackAcquisition(ImageXpressPlateAcquisition):
-    """Image stack acquisition with a Molecular Devices ImageXpress Micro
-    Confocal system.
+class MixedAcquisition(StackAcquisition):
+    """Image stack acquisition with Projectsion acquired with a Molecular
+    Devices ImageXpress Micro Confocal system.
 
     MIP-2P-2sub-Stack --> {name} [Optional]
     └── 2023-02-21 --> {date}
         └── 1334 --> {acquisition id}
+            ├── Projection-Mix_E07_s1_w1E94C24BD-45E4-450A-9919-257C714278F7.tif
+            ├── Projection-Mix_E07_s1_w1_thumb4BFD4018-E675-475E-B5AB-2E959E6B6DA1.tif
+            ├── ...
+            ├── Projection-Mix_E08_s2_w3CCE83D85-0912-429E-9F18-716A085BB5BC.tif
+            ├── Projection-Mix_E08_s2_w3_thumb4D88636E-181E-4AF6-BC53-E7A435959C8F.tif
             ├── ZStep_1
             │   ├── Projection-Mix_E07_s1_w1E78EB128-BD0D-4D94-A6AD-3FF28BB1B105.tif
             │   ├── Projection-Mix_E07_s1_w1_thumb187DE64B-038A-4671-BF6B-683721723769.tif
@@ -37,58 +40,45 @@ class StackAcquisition(ImageXpressPlateAcquisition):
     The *_thumb*.tif files, used by Molecular Devices as preview, are ignored.
     """
 
-    _z_spacing: float = None
-
     def __init__(
         self,
         acquisition_dir: Union[Path, str],
         alignment: TileAlignmentOptions,
-        background_correction_matrices: Optional[dict[str, Union[Path, str]]] = None,
-        illumination_correction_matrices: Optional[dict[str, Union[Path, str]]] = None,
+        background_correction_matrix: Optional[dict[str, NDArray]] = None,
+        illumination_correction_matrix: Optional[NDArray] = None,
     ):
         super().__init__(
             acquisition_dir=acquisition_dir,
             alignment=alignment,
-            background_correction_matrices=background_correction_matrices,
-            illumination_correction_matrices=illumination_correction_matrices,
+            background_correction_matrices=background_correction_matrix,
+            illumination_correction_matrices=illumination_correction_matrix,
         )
 
     def _parse_files(self) -> pd.DataFrame:
         files = super()._parse_files()
-        self._z_spacing = self._compute_z_spacing(files)
-        return files
+        return self._filter_mips(files)
+
+    def _filter_mips(self, files: pd.DataFrame) -> pd.DataFrame:
+        """Remove MIP files if the whole stack was acquired."""
+        _files = files.copy()
+        for ch in _files["channel"].unique():
+            channel_files = _files[_files["channel"] == ch]
+            z_positions = channel_files["z"].unique()
+            has_mip = None in z_positions
+            has_stack = len(z_positions) > 1
+            if has_mip and has_stack:
+                _files.drop(
+                    _files[(_files["channel"] == ch) & (_files["z"].isna())].index,
+                    inplace=True,
+                )
+        return _files
 
     def _get_root_re(self) -> re.Pattern:
         return re.compile(
-            r".*[\/\\](?P<date>\d{4}-\d{2}-\d{2})[\/\\](?P<acq_id>\d+)(?:[\/\\]ZStep_(?P<z>\d+))"
+            r".*[\/\\](?P<date>\d{4}-\d{2}-\d{2})[\/\\](?P<acq_id>\d+)(?:[\/\\]ZStep_(?P<z>\d+))?.*"
         )
 
     def _get_filename_re(self) -> re.Pattern:
         return re.compile(
             r"(?P<name>.*)_(?P<well>[A-Z]+\d{2})_?(?P<field>s\d+)?_?(?P<channel>w[1-9]{1})?(?!_thumb)(?P<md_id>[0-9A-F]{8}-[0-9A-F]{4}-[0-9A-F]{4}-[0-9A-F]{4}-[0-9A-F]{12})(?P<ext>.tif)"
         )
-
-    def _get_z_spacing(self) -> Optional[float]:
-        return self._z_spacing
-
-    def _compute_z_spacing(self, files: pd.DataFrame) -> Optional[float]:
-        assert "z" in files.columns, "No z column in files DataFrame."
-        channel_with_stack = np.sort(files[files["z"] == "2"]["channel"].unique())[0]
-        subset = files[files["channel"] == channel_with_stack]
-        subset = subset[subset["well"] == np.sort(subset["well"].unique())[0]]
-        subset = subset[subset["field"] == np.sort(subset["field"].unique())[0]]
-
-        plane_positions = []
-
-        for i, row in subset.iterrows():
-            file = row["path"]
-            if "z" in row.keys() and row["z"] is not None:
-                metadata = load_metaseries_tiff_metadata(file)
-                z_position = metadata["stage-position-z"]
-                plane_positions.append(z_position)
-
-        plane_positions = np.array(sorted(plane_positions), dtype=np.float32)
-
-        precision = -Decimal(str(plane_positions[0])).as_tuple().exponent
-        z_step = np.round(np.mean(np.diff(plane_positions)), decimals=precision)
-        return z_step
