+++ src/faim_ipa/hcs/imagexpress/SinglePlaneAcquisition.py
@@ -2,8 +2,8 @@ import re
 from pathlib import Path
 from typing import Optional, Union
 
-from faim_hcs.hcs.acquisition import TileAlignmentOptions
-from faim_hcs.hcs.imagexpress import ImageXpressPlateAcquisition
+from faim_ipa.hcs.acquisition import TileAlignmentOptions
+from faim_ipa.hcs.imagexpress import ImageXpressPlateAcquisition
 
 
 class SinglePlaneAcquisition(ImageXpressPlateAcquisition):
