+++ src/faim_ipa/io/ChannelMetadata.py
