+++ src/faim_ipa/stitching/BoundingBox5D.py
