+++ src/faim_ipa/stitching/Tile.py
