+++ src/faim_ipa/stitching/__init__.py
