+++ tests/hcs/cellvoyager/test_CellVoyagerWellAcquisition.py
@@ -5,11 +5,11 @@ from pathlib import Path
 import pandas as pd
 import pytest
 
-from faim_hcs.hcs.acquisition import TileAlignmentOptions
-from faim_hcs.hcs.cellvoyager.CellVoyagerWellAcquisition import (
+from faim_ipa.hcs.acquisition import TileAlignmentOptions
+from faim_ipa.hcs.cellvoyager.CellVoyagerWellAcquisition import (
     CellVoyagerWellAcquisition,
 )
-from faim_hcs.hcs.cellvoyager.StackedTile import StackedTile
+from faim_ipa.hcs.cellvoyager.StackedTile import StackedTile
 
 
 @pytest.fixture
