+++ tox.ini
@@ -34,4 +34,4 @@ deps =
     pytest  # https://docs.pytest.org/en/latest/contents.html
     pytest-cov  # https://pytest-cov.readthedocs.io/en/latest/
     # you can remove these if you don't use them
-commands = pytest -v --color=yes --cov=faim_hcs --cov-report=xml
+commands = pytest -v --color=yes --cov=faim_ipa --cov-report=xml
